Move shared test helpers into utils.ts
